backend/requirements.txt
pandas==1.5.3
pyodbc==5.0.1
python-dotenv==0.21.1
azure-functions==1.18.0
